requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django==5.1.6
django-treebeard==4.7.1
Faker==36.1.1
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
sqlparse==0.5.3
tzdata==2025.1
